display-commonsys-intf: Fix a build error from new tag
hardware/qcom-caf/sm8150/display/display-commonsys-intf/services/config/src/device_impl.cpp:105:63: error: format specifies type 'unsigned long' but the argument has type 'uint64_t' (aka 'unsigned long long') [-Werror,-Wformat]
  ALOGI("Register client id: %lu name: %s device client: %p", client_handle, client_name.c_str(),
                             ~~~                              ^~~~~~~~~~~~~
                                 ^~~~~~~~~~~
hardware/qcom-caf/sm8150/display/display-commonsys-intf/services/config/src/device_impl.cpp:121:41: error: format specifies type 'unsigned long' but the argument has type 'uint64_t' (aka 'unsigned long long') [-Werror,-Wformat]

+++ display-commonsys-intf/services/config/src/device_impl.cpp
@@ -102,7 +102,7 @@ Return<void> DeviceImpl::registerClient(const hidl_string &client_name,
   device_client->SetDeviceConfigIntf(intf);
 
   std::lock_guard<std::mutex> lock(death_service_mutex_);
-  ALOGI("Register client id: %lu name: %s device client: %p", client_handle, client_name.c_str(),
+  ALOGI("Register client id: %lu name: %s device client: %p", (long)client_handle, client_name.c_str(),
         device_client.get());
   display_config_map_.emplace(std::make_pair(client_handle, device_client));
   _hidl_cb(error, client_handle);
@@ -118,7 +118,7 @@ void DeviceImpl::serviceDied(uint64_t client_handle,
     ConfigInterface *intf = client->GetDeviceConfigIntf();
     intf_->UnRegisterClientContext(intf);
     client.reset();
-    ALOGW("Client id:%lu service died", client_handle);
+    ALOGW("Client id:%lu service died", (long)client_handle);
     display_config_map_.erase(itr);
   }
 }
